requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# WebSocket support
python-socketio==5.11.0
websockets==13.1

# Database
sqlalchemy==2.0.36
alembic==1.13.3
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL support

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.12

# AI Integration - Google Gemini
google-generativeai==0.8.3

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# Utilities
aiosqlite==0.20.0  # Async SQLite support
